Mediante autenticación, cada tipo de usuario ve las opciones del Dashboard en base a su rol

+++ app/Http/Controllers/UserController.php
@@ -5,14 +5,17 @@ namespace App\Http\Controllers;
 use App\Http\Requests\UserRequest;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class UserController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
     public function index(Request $request)
     {
+        $this->authorize('user-list');
         $texto = $request->input("texto");
          $registros = User::where("name",'like','%'.$texto.'%')
          ->orWhere('email','like','%'.$texto.'%')
@@ -26,6 +29,7 @@ class UserController extends Controller
      */
     public function create()
     {
+        $this->authorize('user-create');
         return view('usuario.action');
     }
 
@@ -34,6 +38,7 @@ class UserController extends Controller
      */
     public function store(UserRequest $request)
     {
+        $this->authorize('user-create');
         $registro=new User();
         $registro->name=$request->input('name');
         $registro->email=$request->input('email');
@@ -56,6 +61,7 @@ class UserController extends Controller
      */
     public function edit($id)
     {
+        $this->authorize('user-edit');
         $registro = User::findOrFail($id);
 
         return view('usuario.action', compact('registro'));
@@ -66,6 +72,7 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $this->authorize('user-edit');
         $registro=User::findOrFail($id);
         $registro->name=$request->input('name');
         $registro->email=$request->input('email');
@@ -80,6 +87,7 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('user-delte');
         $registro=User::findOrFail($id);
         $registro->delete();
         return redirect()->route('usuarios.index')->with('mensaje', $registro->name. ' <Eliminado Satisfactoriamente.');
